Add files via upload
Problem Statement :

The issuers have a set of model/s and rule/s in place which take information regarding their current financial standing, previous credit history and some other variables as input and output a metric which gives a measure of the risk that the issuer will potentially take on issuing the loan. The measure is generally in the form of a probability and is the risk that the person will default on their loan (called the probability of default) in the future.
Based on the amount of risk that the issuer is willing to take (plus some other factors) they decide on a cutoff of that score and use it to take a decision regarding whether to pass the loan or not. This is a way of managing credit risk. The whole process collectively is referred to as underwriting.

Solution :

•	Data Cleaning Processing
•	Checking the variable type count in data
•	Identifying the all unique values
•	Dealing with Unique, or same value columns
•	Counting the frequency of variables to check the biasness
•	Dealing with all biased column.
•	Drop all the columns having more than 75% of NA values
•	For remaining columns Imputation according to data type
•	Get numerical and categorial column separate
•	Outlier analysis
•	Discover outliers with mathematical function
•	Rescaling Data
•	Check for Correlation with Pearson Correlation to get all the Numerical Columns first with below code
•	Split into test & train data 
	
FITTING MODELS TO DATA

•	Logistic regression 
•	Random forest